LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/font-relative-units-dynamic-expected.txt


PASS em relative before change
PASS rem relative before change
PASS ex relative before change
PASS ch relative before change
FAIL em relative after change assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
FAIL rem relative after change assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
FAIL ex relative after change assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
FAIL ch relative after change assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
